Add op ones and move new op to op/training/
