+++ src/relax/op/tensor/binary.cc
@@ -103,20 +103,6 @@ Type InferTypeBinaryBroadcast(const Call& call, DiagnosticContext diag_ctx) {
   return DynTensorType(output_ndim, output_dtype);
 }
 
-Optional<Expr> InferShapeCollapseSumLike(const Call& call, DiagnosticContext diag_ctx) {
-  if (call->args.size() != 2) {
-    diag_ctx.EmitFatal(Diagnostic::Error(call->span) << "Matmul op should have 2 arguments");
-  }
-  return call->args[1]->shape();
-}
-
-Type InferTypeCollapseSumLike(const Call& call, DiagnosticContext diag_ctx) {
-  if (call->args.size() != 2) {
-    diag_ctx.EmitFatal(Diagnostic::Error(call->span) << "Matmul op should have 2 arguments");
-  }
-  return call->args[1]->checked_type();
-}
-
 RELAX_REGISTER_BINARY_BROADCAST_OP("add")
     .describe("Elementwise add with broadcasting")
     .set_support_level(1);
@@ -128,10 +114,6 @@ RELAX_REGISTER_BINARY_BROADCAST_OP("sub")
 RELAX_REGISTER_BINARY_BROADCAST_OP("multiply")
     .describe("Elementwise multiply with broadcasting")
     .set_support_level(1);
-
-
-RELAX_REGISTER_BINARY_OP_BASE("collapse_sum_like", InferShapeCollapseSumLike, InferTypeCollapseSumLike);
-
 
 }  // namespace relax
 }  // namespace tvm
